Add IndexProperties to report which proximity index results are from.
Add IndexProperties to record the ID of the index (e.g. RTree) results
are coming from and their properties, e.g. the part of the tree they
are for (e.g. which server they are replicated from) and whether they
are for dynamic objects (e.g. so we can enable optimizations in the
replicated index).