Add list of indexes to Loc updates.
The list of indexes indicate which indexes the object is currently
used by for the update, so it's a repeated field. The field isn't
required for old-style simple querying. It's only necessary to allow
clients to know which index(es) the update needs to be applied to,
including deferred application for 'orphan' loc updates that arrive
before the the prox update that adds the object.